pyproject.toml
[build-system]
requires = ["poetry-core>=1.4.2"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "start-django"
version = "0.0.1"
description = "Django + fly.io, styled by Tailwind, htmx-ed"
authors = ["Marcelino G. Veloso III <mars@veloso.one>"]

[tool.poetry.dependencies]
python = "^3.11"                    # base
django = "^4.2"                     # base
dj-database-url = "^0.5.0"          # base: db
psycopg = "^3.1.8"                  # base: postgres; remove if sqlite-only
django-fragments = "^0.1.8"         # custom: template tags, bs4, markdown
cloudflare-images = "^0.0.2"        # custom: images-only storage class
django-allauth = "^0.54.0"          # common: auth + social
django-widget-tweaks = "^1.4.12"    # common: form widget managemnt in html
python-postmark = "^0.6.0"          # service: transactional email
huey = "^2.4.5"                     # service: background worker
redis = "^4.5.4"                    # service: message broker / cache
pillow = "^9.5.0"                   # static: imagefield
Brotli = "^1.0.9"                   # static: whitenoise staticfiles
django-compressor = "^4.3.1"        # static: staticfiles compression
whitenoise = "^6.4.0"               # static: unused fly.io; usable gunicorn local
environs = "^9.5.0"                 # tooling: dotenv
python-slugify = "^8.0.1"           # tooling: id/slugs
pyyaml = "^6.0"                     # tooling: used in fixtures
python-dateutil = "^2.8.2"          # tooling: date parsing
django-debug-toolbar = "^4.0.0"     # dev: debug
django-extensions = "^3.2.1"        # dev: multi-purpose
rich = "^13.3.5"                    # dev: shell formatting / colors
gunicorn = "^20.1.0"                # prod: server
sentry-sdk = "^1.21.0"              # prod: logging

[tool.poetry.group.dev.dependencies]
pytest = "^7.2"
pytest-django = "^4.5.1"
pytest-cov = "^2.12.1"
pre-commit = "^2.21"
mkdocs = "^1.4.2"
mkdocstrings = { extras = ["python"], version = "^0.20.0" }
mkdocs-material = "^9.1"
djhtml = "3.0.6"
ipykernel = "^6.22.0"
werkzeug = "^2.2"

[tool.pytest.ini_options]
minversion = "7.2"
pythonpath = "src" # enable pytest in root directly
addopts = "-ra -q --ds=config.settings --doctest-modules --cov"
filterwarnings = [
  "ignore::DeprecationWarning", # DeprecationWarning: pkg_resources is deprecated as an API
  "ignore::django.utils.deprecation.RemovedInDjango51Warning" # GET_STORAGE_CLASS_DEPRECATED_MSG
]
testpaths = ["tests", "src"]

[tool.ruff]
ignore = ["F401", "F403"]
fixable = ["F", "E", "W", "I001"]
select = ["F", "E", "W", "I001"]

[tool.ruff.per-file-ignores]
"src/config/settings/__init__.py" = ["F405", "E501"]
"src/config/settings/_auth.py" = ["F405", "E501"]
"src/config/settings/_settings.py" = ["E501"]

[tool.djlint]
profile="django"
